cambios landing y carrito

+++ src/components/ProductCards/productCards.jsx
@@ -1,11 +1,14 @@
 'use client';
+import { removeFromCart, addToCart } from '@/redux/Cart/cartSlice';
 import { useRouter } from 'next/navigation';
+import { useDispatch } from 'react-redux';
 
-export default function ProductCard({ product, admin, setSelectProduct, setIsModalOpen, handleDelete }) {
+export default function ProductCard({ product, admin, setSelectProduct, setIsModalOpen, handleDelete, cart }) {
     const router = useRouter();
+    const dispatch = useDispatch();
 
     const handleCardClick = () => {
-        if (!admin) {  // Si no es admin, se realiza la redirección
+        if (!admin && !cart) {  // Si no es admin, se realiza la redirección
             const url = `/Home/Product/${product.id}`;
             router.push(url);
         }
@@ -18,13 +21,26 @@ export default function ProductCard({ product, admin, setSelectProduct, setIsMod
     };
 
     const handleDeleteClick = (e) => {
-        handleDelete(product.id); // Llamar a la función de eliminar el producto
+        if (admin) {
+            handleDelete(product.id);
+        } // Llamar a la función de eliminar el producto
+
+        if (cart) {
+            dispatch(removeFromCart(product))
+        }
+    };
+
+
+    const handleAddToCart = (e) => {
+        if (cart) {
+            dispatch(addToCart(product))
+        }
     };
 
     return (
         <div
             onClick={handleCardClick} // Si no es admin, redirige; si es admin, no hace nada
-            className="flex cursor-pointer flex-col border p-4 rounded-lg shadow-lg hover:shadow-2xl transition-shadow duration-300 max-w-[15rem] h-80 sm:w-1/2 md:w-1/3 lg:w-1/4"
+            className="flex cursor-pointer flex-col border p-2 md:p-4 rounded-lg shadow-lg hover:shadow-2xl transition-shadow duration-300 w-60 h-80 "
         >
             <div className="w-full h-52 overflow-hidden rounded-lg mb-4">
                 <img
@@ -45,6 +61,25 @@ export default function ProductCard({ product, admin, setSelectProduct, setIsMod
                     >
                         Editar
                     </button>
+
+                    <button
+                        onClick={handleDeleteClick}
+                        className="bg-red-500 text-white px-4 py-2 rounded-lg hover:bg-red-600 transition"
+                    >
+                        Eliminar
+                    </button>
+                </div>
+            )}
+            {/* Boton de carrito: Eliminar y agregar */}
+            {cart && (
+                <div className="flex space-x-4 mt-4">
+                    <button
+                        onClick={handleAddToCart}
+                        className="bg-blue-500 text-white px-4 py-2 rounded-lg hover:bg-blue-600 transition"
+                    >
+                        Agregar
+                    </button>
+
                     <button
                         onClick={handleDeleteClick}
                         className="bg-red-500 text-white px-4 py-2 rounded-lg hover:bg-red-600 transition"
